refactor: isolate compaction planning
